refactor: 전시회 목록 GET 요청 api와 훅으로 분리
