+++ src/components/exhibition-list.tsx
@@ -1,21 +1,8 @@
-import { useQuery } from '@tanstack/react-query';
-
 import ExhibitionCard from '@/components/exhibition-card';
-import type { Exhibition } from '@/types/exhibition-type';
+import { useExhibition } from '@/hooks/useExhibition';
 
 export default function ExhibitionList() {
-  const { data: exhibitions } = useQuery({
-    queryKey: ['exhibitions'],
-    queryFn: async (): Promise<Array<Exhibition>> => {
-      const response = await fetch('/exhibitions', {
-        method: 'GET',
-      });
-      if (!response.ok) throw new Error('GET /exhibitions');
-      const exhibitions = response.json();
-
-      return exhibitions;
-    },
-  });
+  const exhibitions = useExhibition();
 
   // TODO: handle when length === 0
   return (
